flask_app/netology_app/requirements.txt
flask==0.12
redis==2.10.5
pymongo==3.6.1
psycopg2==2.7.1
numpy==1.13.1
scipy==0.19.1
lz4==0.10.1
msgpack-numpy==0.4.1
msgpack-python==0.4.8